+++ posthog/api/decide.py
@@ -15,7 +15,7 @@ from posthog.models import Team, User
 from posthog.models.feature_flag import get_overridden_feature_flags
 from posthog.utils import cors_response, load_data_from_request
 
-from .capture import _get_project_id
+from .utils import get_project_id_from_request
 
 
 def on_permitted_domain(team: Team, request: HttpRequest) -> bool:
@@ -96,7 +96,7 @@ def get_decide(request: HttpRequest):
         token, _ = get_token(data, request)
         team = Team.objects.get_team_from_token(token)
         if team is None and token:
-            project_id = _get_project_id(data, request)
+            project_id = get_project_id_from_request(data, request)
 
             if not project_id:
                 return cors_response(
